Support boolean subqueries in addition to boolean expressions in filters
There were couple of issues in filter expressions this commit fixes:

1. Filters allowed all kinds of expressions, not just boolean expressions.
2. Filters didn't allow subqueries that select a boolean value.

This commit can be considered a breaking change since people will need
to add explicit types for raw sql filters if they didn't have them already.
On the other hand this commit fixes a typing bug, so it's up to interpretation.