Update isPlainObject to exclude objects with constructors
If you are using a custom type parser (via `pg.types.setTypeParser`)
that returns a custom class, the camel case plugin will incorrectly
serialize that value to a literal object, stripping it of its prototype.

This is problematic if you rely on the ability to call certain methods
specified by the custom class' prototype.