Fixes JsonAgg type definition
Updates the type definition for JsonAgg aggregate function to correctly infer the element type of the resulting array.

The previous type definition used `Selectable` which was too broad. The updated type definition now uses `SelectType` to extract the correct type and `ShallowDehydrateValue` to remove `undefined` from the type.
The change is accompanied by a fix in the typings tests.